refactor(drawio): 收敛编辑器会话状态
