+++ src/views/workspace/drawio/index.tsx
@@ -15,18 +15,17 @@ import {
   useResourceHostLayoutConfig,
   type ResourceHostLayoutConfig,
 } from '@/views/workspace/ResourceHostContext';
-import { Button, toast } from '@heroui/react';
-import { useEventListener, useRequest, useUnmount } from 'ahooks';
+import { Button } from '@heroui/react';
+import { useRequest } from 'ahooks';
 import { History, Save } from 'lucide-react';
-import { useRef, useState, type ReactNode } from 'react';
+import { useState, type ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
+import { useDrawioEditorSession } from './_hooks/useDrawioEditorSession';
 import {
   buildDrawioUrl,
   decodeBase64Utf8,
-  extractDrawioPlainText,
   readDrawioEmbedOrigin,
-  readDrawioMessage,
   type DrawioSaveState,
   type WisePenTheme,
 } from './drawioProtocol';
@@ -214,16 +213,7 @@ function DrawioViewConnected({ resourceId, data, onRefreshDrawioInfo }: DrawioVi
   const { noteInfoDisplay, snapshot, initialXml } = data;
   const noteService = useNoteService();
   const userService = useUserService();
-  const iframeRef = useRef<HTMLIFrameElement>(null);
   const initialVersion = Math.max(noteInfoDisplay.version ?? 0, snapshot.version ?? 0);
-  const currentVersionRef = useRef(initialVersion);
-  const lastSavedXmlRef = useRef(initialXml);
-  const exportTimerRef = useRef<number | null>(null);
-  const pendingExportForSaveRef = useRef(false);
-  const [currentVersion, setCurrentVersion] = useState(initialVersion);
-  const [saveState, setSaveState] = useState<DrawioSaveState>('saved');
-  const [editorReady, setEditorReady] = useState(false);
-  const [editorLoaded, setEditorLoaded] = useState(false);
   const [versionOpen, setVersionOpen] = useState(false);
   const canEdit = noteInfoDisplay.canCollaborativeEdit;
   const canViewVersions = Boolean(noteInfoDisplay.ownerId);
@@ -236,135 +226,20 @@ function DrawioViewConnected({ resourceId, data, onRefreshDrawioInfo }: DrawioVi
     colorScheme: readWisePenColorScheme(),
   });
   const drawioOrigin = readDrawioEmbedOrigin(DRAWIO_EMBED_URL);
+  const { iframeRef, currentVersion, saveState, editorReady, editorLoaded, requestSave } =
+    useDrawioEditorSession({
+      canEdit,
+      drawioOrigin,
+      initialVersion,
+      initialXml,
+      noteService,
+      resourceId,
+    });
 
   const { data: currentUser } = useRequest(() => userService.getUserInfo(), {
     ready: Boolean(noteInfoDisplay.ownerId),
     refreshDeps: [noteInfoDisplay.ownerId],
   });
-
-  const postToEditor = (message: Record<string, unknown>) => {
-    iframeRef.current?.contentWindow?.postMessage(JSON.stringify(message), drawioOrigin);
-  };
-
-  const clearExportTimer = () => {
-    if (exportTimerRef.current !== null) {
-      window.clearTimeout(exportTimerRef.current);
-      exportTimerRef.current = null;
-    }
-  };
-
-  const persistXml = async (xml: string) => {
-    if (!canEdit) {
-      toast.danger(t('drawio.noEditPermission'));
-      return;
-    }
-
-    const nextVersion = currentVersionRef.current + 1;
-    setSaveState('saving');
-    try {
-      await noteService.saveDrawIoSnapshot({
-        resourceId,
-        version: nextVersion,
-        xml,
-        plainText: extractDrawioPlainText(xml),
-      });
-      currentVersionRef.current = nextVersion;
-      lastSavedXmlRef.current = xml;
-      setCurrentVersion(nextVersion);
-      setSaveState('saved');
-      postToEditor({ action: 'status', message: t('drawio.status.saved'), modified: false });
-      toast.success(t('drawio.status.saved'));
-    } catch (err) {
-      setSaveState('failed');
-      postToEditor({ action: 'status', message: t('drawio.status.failed'), modified: true });
-      toast.danger(parseErrorMessage(err));
-    }
-  };
-
-  const requestEditorExport = () => {
-    if (!canEdit) {
-      toast.danger(t('drawio.noEditPermission'));
-      return;
-    }
-    if (!editorLoaded) {
-      toast.info(t('drawio.editorNotReady'));
-      return;
-    }
-    if (saveState === 'saved') {
-      toast.info(t('drawio.alreadySaved'));
-      return;
-    }
-
-    pendingExportForSaveRef.current = true;
-    setSaveState('saving');
-    postToEditor({ action: 'export', format: 'xml' });
-    clearExportTimer();
-    exportTimerRef.current = window.setTimeout(() => {
-      pendingExportForSaveRef.current = false;
-      setSaveState('failed');
-      toast.danger(t('drawio.saveRetry'));
-    }, 10000);
-  };
-
-  const handleMessage = (event: MessageEvent) => {
-    if (event.origin !== drawioOrigin) return;
-    if (event.source !== iframeRef.current?.contentWindow) return;
-
-    const message = readDrawioMessage(event.data);
-    if (!message?.event) return;
-
-    if (message.event === 'init') {
-      setEditorReady(true);
-      postToEditor({
-        action: 'load',
-        autosave: canEdit ? 1 : 0,
-        modified: false,
-        noExitBtn: 1,
-        noSaveBtn: canEdit ? 0 : 1,
-        saveAndExit: 0,
-        xml: initialXml,
-      });
-      return;
-    }
-
-    if (message.event === 'load') {
-      setEditorLoaded(true);
-      setSaveState('saved');
-      return;
-    }
-
-    if (message.event === 'autosave' && canEdit && typeof message.xml === 'string') {
-      if (message.xml !== lastSavedXmlRef.current && saveState !== 'saving') {
-        setSaveState('dirty');
-      }
-      return;
-    }
-
-    if (message.event === 'save' && typeof message.xml === 'string') {
-      void persistXml(message.xml);
-      return;
-    }
-
-    if (message.event === 'export' && pendingExportForSaveRef.current) {
-      pendingExportForSaveRef.current = false;
-      clearExportTimer();
-      if (typeof message.xml === 'string') {
-        void persistXml(message.xml);
-      } else {
-        setSaveState('failed');
-        toast.danger(t('drawio.saveRetry'));
-      }
-      return;
-    }
-
-    if (message.event === 'error') {
-      toast.danger(message.message || t('drawio.loadFailed'));
-    }
-  };
-
-  useEventListener('message', handleMessage);
-
-  useUnmount(clearExportTimer);
 
   const {
     data: versions,
@@ -405,7 +280,7 @@ function DrawioViewConnected({ resourceId, data, onRefreshDrawioInfo }: DrawioVi
           size="sm"
           variant="primary"
           isDisabled={!editorLoaded || saveState === 'saved' || saveState === 'saving'}
-          onPress={requestEditorExport}
+          onPress={requestSave}
           aria-label={t('actions.save', { ns: 'common' })}
         >
           <Save size={16} />
